Extract CellWidget from PlotGridTabs (no behavior change)
This feature branch grew the per-cell view substantially (titlebar, rename,
per-cell add-layer, toolbar toggle, freshness/lag pill, per-layer time
labels). That work had been accreting onto PlotGridTabs as methods plus five
parallel per-cell/per-layer dicts that every lifecycle path had to hand-sync.

Consolidate the per-cell view into a new CellWidget class (widgets/cell.py)
that owns its titlebar, layer toolbars, content, autoscale controller,
freshness pane, and layer-time panes, exposing update_freshness(),
update_layer_time(), and dispose(). PlotGridTabs now holds a single
dict[CellId, CellWidget] (plus the global _session_layers version registry,
which belongs to the poll loop), so the scattered cleanup loops collapse to
delegated calls.

The freshness/layer-time render helpers move with it: their only caller is
CellWidget, so the formatter and its consumer now live together.

Pure refactor — no behavior change; the full dashboard suite passes
unchanged.
